Added new features :+1:
Now a new folder will be created.
There will be a index.html file ( the main file of the report ),
And ( for now ) a folder called: Classes.

This Classes folder now has a html file for everyone of the classes in
the report and in them there is a Code Coverage Color Report by lines. I
think it's nice for now.